+++ backend/celery_app/__init__.py
@@ -1,3 +1,0 @@
-from celery_app.celery import app as celery
-
-__all__ = ["celery"]